server/requirements.txt
grpcio>=1.60.0
grpcio-tools==1.71.0
# Real inference: load converted Keras model
tensorflow>=2.15.0
tensorflowjs>=4.0.0
Pillow>=10.0.0
numpy>=1.24.0
protobuf==5.29.6
